+++ oboron/tests/feature_gate_tests.rs
@@ -17,7 +17,7 @@ mod bytes_keys_enabled {
         {
             let _ = AasvC32::from_bytes(&key);
             let _ = Ob::from_bytes("aasv.c32", &key);
-            let _ = ObMulti::from_bytes(&key);
+            let _ = Omnib::from_bytes(&key);
         }
     }
 
@@ -48,7 +48,7 @@ mod hex_keys_enabled {
         {
             let _ = AasvC32::from_hex_key(&key);
             let _ = Ob::from_hex_key("aasv.c32", &key);
-            let _ = ObMulti::from_key_hex(&key);
+            let _ = Omnib::from_key_hex(&key);
         }
     }
 
@@ -77,7 +77,7 @@ mod keyless_enabled {
         {
             let _ = AasvC32::new_keyless();
             let _ = Ob::new_keyless("aasv.c32");
-            let _ = ObMulti::new_keyless();
+            let _ = Omnib::new_keyless();
         }
     }